Code clean-up
It's been a few years since I've been on this site. I need to make
things right.

This is some cleaup (mostly PEP-8. a little Typing) that I wanted to
apply to the code before taking the PR that fixes the sympy change.

Adding the PR 4 (naari3's) to the mix. Will have to continue on
and work a bit the the sp numbers too.